+++ src/config/buttons/types/PeriodShiftButton.ts
@@ -2,19 +2,21 @@ import {ButtonInterface} from "../ButtonInterface";
 import {ButtonBaseConfig} from "../ButtonBaseConfig";
 import {CardConfigButton} from "../../../card/CardConfigButton";
 import {ButtonColors} from "../ButtonColors";
-import {ConfigCheckUtils} from "../../ConfigCheckUtils";
+import {PeriodCodeHelper} from "../../../util/PeriodCodeHelper";
 
 export class PeriodShiftButton
     implements ButtonInterface {
 
     constructor(
         public readonly baseConfig: ButtonBaseConfig,
-        public readonly hours: number) {
+        public readonly shiftPeriod: string) {
     }
 
     static fromConfig(button: CardConfigButton, defaultColors: ButtonColors): PeriodShiftButton {
         const baseConfig = ButtonBaseConfig.fromConfig(button, defaultColors);
-        const hours = ConfigCheckUtils.checkNumber("hours", button.hours);
-        return new PeriodShiftButton(baseConfig, hours);
+        if (PeriodCodeHelper.check("shift_period", button.shift_period)) {
+            return new PeriodShiftButton(baseConfig, button.shift_period);
+        }
+        throw new Error("WindRoseCard: ")
     }
 }